+++ src/entity/static_mesh.hpp
@@ -1,4 +1,4 @@
-#include "../entity/mesh.hpp"
+#include "mesh.hpp"
 #include "../gui/meshloader.hpp"
 
 #include <GL/glew.h>
